Added some error handling with exceptions. Moved the task and command files into their respective folders.
