+++ src/main/java/duke/commands/DoneCommand.java
@@ -1,4 +1,7 @@
-package duke;
+package duke.commands;
+
+import duke.Duke;
+import duke.tasks.Task;
 
 public class DoneCommand extends Command {
     private static final String DONE_TASK_STRING = "Nice! I've marked this task as done:\n";
